moving jobs from autoplex/data/phonons/jobs.py to autoplex/data/common/jobs.py
